Move ensuredirectory from output to utils
