+++ output.cc
@@ -22,7 +22,6 @@ struct STAT{                                           // Stores statistical inf
 	double ESS;                                          // The estimated sample size
 };
 
-static void ensuredirectory(const string &path);
 STAT getstat(vector <double> &vec);
 
 ofstream trace, traceLi;
@@ -525,17 +524,4 @@ STAT getstat(vector <double> &vec)
 	}
 		
 	return stat;
-}
-
-/// Create a directory if it doesn't already exist
-static void ensuredirectory(const string &path) 
-{
-	struct stat st;
-	if (stat(path.c_str(), &st) == -1)
-	{
-		// Directory not found
-		int ret = mkdir(path.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
-		if (ret == -1)
-			emsg("Error creating directory '"+path+"'");
-	}
 }
